FINISHED - 观察者模式 event-listener
