+++ morphiademo/src/main/java/com/ninggc/demo/morphiademo/morphia/dao/MorphiaQsAnswerDAO.java
@@ -1,0 +1,7 @@
+package com.ninggc.demo.morphiademo.morphia.dao;
+
+import com.ninggc.demo.morphiademo.morphia.dao.base.MorphiaBase;
+import com.ninggc.demo.morphiademo.morphia.domain.MorphiaQsAnswerPO;
+
+public interface MorphiaQsAnswerDAO extends MorphiaBase<MorphiaQsAnswerPO> {
+}
