+++ morphiademo/src/main/java/com/ninggc/demo/morphiademo/morphia/dao/impl/MorphiaQsAnswerDAOImpl.java
@@ -1,0 +1,13 @@
+package com.ninggc.demo.morphiademo.morphia.dao.impl;
+
+import com.ninggc.demo.morphiademo.morphia.dao.MorphiaQsAnswerDAO;
+import com.ninggc.demo.morphiademo.morphia.dao.base.impl.MorphiaBaseImpl;
+import com.ninggc.demo.morphiademo.morphia.domain.MorphiaQsAnswerPO;
+import org.springframework.stereotype.Service;
+
+/**
+ * @author ninggc
+ */
+@Service
+public class MorphiaQsAnswerDAOImpl extends MorphiaBaseImpl<MorphiaQsAnswerPO> implements MorphiaQsAnswerDAO {
+}
