+++ morphiademo/src/main/java/com/ninggc/demo/morphiademo/morphia/exception/NotFoundException.java
@@ -1,0 +1,4 @@
+package com.ninggc.demo.morphiademo.morphia.exception;
+
+public class NotFoundException extends Exception {
+}
