+++ morphiademo/src/main/java/com/ninggc/demo/morphiademo/morphia/util/MorphiaMatchMode.java
@@ -1,0 +1,54 @@
+package com.ninggc.demo.morphiademo.morphia.util;
+
+public enum MorphiaMatchMode {
+
+    /**
+     * Match the entire string to the pattern
+     */
+    EXACT {
+        @Override
+        public String toMatchString(String pattern) {
+            return pattern;
+        }
+    },
+
+    /**
+     * Match the start of the string to the pattern
+     */
+    START {
+        @Override
+        public String toMatchString(String pattern) {
+            return pattern + '%';
+        }
+    },
+
+    /**
+     * Match the end of the string to the pattern
+     */
+    END {
+        @Override
+        public String toMatchString(String pattern) {
+            return '%' + pattern;
+        }
+    },
+
+    /**
+     * Match the pattern anywhere in the string
+     */
+    ANYWHERE {
+        @Override
+        public String toMatchString(String pattern) {
+            return '%' + pattern + '%';
+        }
+    };
+
+    /**
+     * Convert the pattern, by appending/prepending "%"
+     *
+     * @param pattern The pattern for convert according to the mode
+     *
+     * @return The converted pattern
+     */
+    public abstract String toMatchString(String pattern);
+
+}
