--- morphiademo/src/main/java/com/ninggc/morphiademo/controller/MorphiaController.java
@@ -1,23 +1,0 @@
-package com.ninggc.morphiademo.controller;
-
-import com.ninggc.morphiademo.morphia.dao.MorphiaQsAnswerDAO;
-import com.ninggc.morphiademo.morphia.domain.MorphiaQsAnswerPO;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.annotation.Resource;
-import java.util.List;
-
-@RestController
-@RequestMapping("/morphia")
-public class MorphiaController {
-    @Resource
-    private MorphiaQsAnswerDAO qsAnswerDAO;
-
-    @GetMapping("/getList")
-    public List<MorphiaQsAnswerPO> getList() {
-        return qsAnswerDAO.findAll(0, 10);
-    }
-}
