--- morphiademo/src/main/java/com/ninggc/morphiademo/morphia/dao/MorphiaQsAnswerDAO.java
@@ -1,7 +1,0 @@
-package com.ninggc.morphiademo.morphia.dao;
-
-import com.ninggc.morphiademo.morphia.dao.base.MorphiaBase;
-import com.ninggc.morphiademo.morphia.domain.MorphiaQsAnswerPO;
-
-public interface MorphiaQsAnswerDAO extends MorphiaBase<MorphiaQsAnswerPO> {
-}
