--- morphiademo/src/main/java/com/ninggc/morphiademo/morphia/dao/impl/MorphiaQsAnswerDAOImpl.java
@@ -1,13 +1,0 @@
-package com.ninggc.morphiademo.morphia.dao.impl;
-
-import com.ninggc.morphiademo.morphia.dao.MorphiaQsAnswerDAO;
-import com.ninggc.morphiademo.morphia.dao.base.impl.MorphiaBaseImpl;
-import com.ninggc.morphiademo.morphia.domain.MorphiaQsAnswerPO;
-import org.springframework.stereotype.Service;
-
-/**
- * @author ninggc
- */
-@Service
-public class MorphiaQsAnswerDAOImpl extends MorphiaBaseImpl<MorphiaQsAnswerPO> implements MorphiaQsAnswerDAO {
-}
