--- morphiademo/src/main/java/com/ninggc/morphiademo/morphia/exception/NotFoundException.java
@@ -1,4 +1,0 @@
-package com.ninggc.morphiademo.morphia.exception;
-
-public class NotFoundException extends Exception {
-}
