--- morphiademo/src/main/java/com/ninggc/morphiademo/morphia/util/MorphiaMatchMode.java
@@ -1,54 +1,0 @@
-package com.ninggc.morphiademo.morphia.util;
-
-public enum MorphiaMatchMode {
-
-    /**
-     * Match the entire string to the pattern
-     */
-    EXACT {
-        @Override
-        public String toMatchString(String pattern) {
-            return pattern;
-        }
-    },
-
-    /**
-     * Match the start of the string to the pattern
-     */
-    START {
-        @Override
-        public String toMatchString(String pattern) {
-            return pattern + '%';
-        }
-    },
-
-    /**
-     * Match the end of the string to the pattern
-     */
-    END {
-        @Override
-        public String toMatchString(String pattern) {
-            return '%' + pattern;
-        }
-    },
-
-    /**
-     * Match the pattern anywhere in the string
-     */
-    ANYWHERE {
-        @Override
-        public String toMatchString(String pattern) {
-            return '%' + pattern + '%';
-        }
-    };
-
-    /**
-     * Convert the pattern, by appending/prepending "%"
-     *
-     * @param pattern The pattern for convert according to the mode
-     *
-     * @return The converted pattern
-     */
-    public abstract String toMatchString(String pattern);
-
-}
