+++ morphiademo/src/test/java/com/ninggc/demo/morphiademo/MorphiademoApplicationTests.java
@@ -1,0 +1,16 @@
+package com.ninggc.demo.morphiademo;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class MorphiademoApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
