--- morphiademo/src/test/java/com/ninggc/morphiademo/MorphiademoApplicationTests.java
@@ -1,16 +1,0 @@
-package com.ninggc.morphiademo;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class MorphiademoApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
