+++ springbootdemo/src/main/java/com/ninggc/demo/DemoApplication.java
@@ -1,22 +1,23 @@
 package com.ninggc.demo;
 
-import com.ninggc.listener.TestEvent;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationListener;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.event.ContextRefreshedEvent;
+import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.support.RequestHandledEvent;
 
-import javax.annotation.Resource;
 import java.util.HashMap;
 import java.util.Map;
 
-@SpringBootApplication(scanBasePackageClasses = {DemoApplication.class})
+@SpringBootApplication
 @RestController
-//@ImportResource(locations = {"classpath:provider.xml"})
-public class DemoApplication implements ApplicationListener<ContextRefreshedEvent> {
+public class DemoApplication implements ApplicationListener<RequestHandledEvent> {
+
+    @Autowired
+    AbstractApplicationContext context;
 
     @GetMapping("")
     public Map<String, Object> test() {
@@ -25,8 +26,19 @@ public class DemoApplication implements ApplicationListener<ContextRefreshedEven
         return map;
     }
 
+    @GetMapping("/beans")
+    public Map<String, Object> beans() {
+        Map<String, Object> map = new HashMap<>();
+        String[] names = context.getBeanDefinitionNames();
+        map.put("name", "spring boot");
+        for (String name : names) {
+            map.put(name, "");
+        }
+        return map;
+    }
+
     @Override
-    public void onApplicationEvent(ContextRefreshedEvent event) {
+    public void onApplicationEvent(RequestHandledEvent event) {
         System.out.println("app start");
     }
