+++ springbootdemo/src/main/java/com/ninggc/demo/service/TestService.java
@@ -1,0 +1,8 @@
+package com.ninggc.demo.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface TestService {
+    void test();
+}
